requirements.txt
pandas>=1.5.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
google-generativeai>=0.3.0
wikipedia>=1.4.0
tqdm>=4.64.0

